+++ ksc-sdk-java-cdn/src/main/java/com/ksc/cdn/model/domain/createdomain/AddDomainRequest.java
@@ -1,4 +1,4 @@
-package com.ksc.cdn.model.domain;
+package com.ksc.cdn.model.domain.createdomain;
 
 import com.ksc.cdn.model.valid.FieldValidate;
 import org.apache.commons.lang3.StringUtils;
