+++ ksc-sdk-java-cdn/src/main/java/com/ksc/cdn/model/domain/domainbase/GetDomainBaseResult.java
@@ -1,4 +1,4 @@
-package com.ksc.cdn.model.domain;
+package com.ksc.cdn.model.domain.domainbase;
 
 /**
  * GetDomainBaseResult
