+++ ksc-sdk-java-cdn/src/main/java/com/ksc/cdn/model/domain/domaincollect/GetCdnDomainsRequest.java
@@ -1,4 +1,4 @@
-package com.ksc.cdn.model.domain;
+package com.ksc.cdn.model.domain.domaincollect;
 
 import org.apache.commons.lang3.StringUtils;
