+++ ksc-sdk-java-cdn/src/main/java/com/ksc/cdn/model/domain/domaindetail/BackOriginHostConfig.java
@@ -1,0 +1,19 @@
+package com.ksc.cdn.model.domain.domaindetail;
+
+/**
+ * BackOriginHostConfig
+ *
+ * @author jiangran@kingsoft.com
+ * @date 2016/11/15
+ */
+public class BackOriginHostConfig {
+    private String BackOriginHost;
+
+    public String getBackOriginHost() {
+        return BackOriginHost;
+    }
+
+    public void setBackOriginHost(String backOriginHost) {
+        BackOriginHost = backOriginHost;
+    }
+}
