+++ ksc-sdk-java-cdn/src/main/java/com/ksc/cdn/model/domain/domaindetail/CacheRuleConfig.java
@@ -1,0 +1,19 @@
+package com.ksc.cdn.model.domain.domaindetail;
+
+/**
+ * CacheRuleConfig
+ *
+ * @author jiangran@kingsoft.com
+ * @date 2016/11/15
+ */
+public class CacheRuleConfig{
+    private CacheRules[] CacheRules;
+
+    public CacheRules[] getCacheRules() {
+        return CacheRules;
+    }
+
+    public void setCacheRules(CacheRules[] cacheRules) {
+        CacheRules = cacheRules;
+    }
+}
