+++ ksc-sdk-java-cdn/src/main/java/com/ksc/cdn/model/domain/domaindetail/HttpHeaderRules.java
@@ -1,0 +1,19 @@
+package com.ksc.cdn.model.domain.domaindetail;
+
+/**
+ * HttpHeaderRules
+ *
+ * @author jiangran@kingsoft.com
+ * @date 2016/11/15
+ */
+public class HttpHeaderRules{
+    private Object[] HttpHeaderRules;
+
+    public Object[] getHttpHeaderRules() {
+        return HttpHeaderRules;
+    }
+
+    public void setHttpHeaderRules(Object[] httpHeaderRules) {
+        HttpHeaderRules = httpHeaderRules;
+    }
+}
