+++ ksc-sdk-java-cdn/src/main/java/com/ksc/cdn/model/domain/domaindetail/HttpHeadersConfig.java
@@ -1,0 +1,11 @@
+package com.ksc.cdn.model.domain.domaindetail;
+
+/**
+ * HttpHeadersConfig
+ *
+ * @author jiangran@kingsoft.com
+ * @date 2016/11/15
+ */
+public class HttpHeadersConfig {
+    private HttpHeaderRules[] HttpHeaderRules;
+}
