+++ ksc-sdk-java-cdn/src/main/java/com/ksc/cdn/model/domain/domaindetail/IpProtectionConfig.java
@@ -1,0 +1,19 @@
+package com.ksc.cdn.model.domain.domaindetail;
+
+/**
+ * IpProtectionConfig
+ *
+ * @author jiangran@kingsoft.com
+ * @date 2016/11/15
+ */
+public class IpProtectionConfig{
+    private String IpList;
+
+    public String getIpList() {
+        return IpList;
+    }
+
+    public void setIpList(String ipList) {
+        IpList = ipList;
+    }
+}
