+++ ksc-sdk-java-cdn/src/main/java/com/ksc/cdn/model/domain/domaindetail/ReferProtectionRequest.java
@@ -1,4 +1,4 @@
-package com.ksc.cdn.model.domain;
+package com.ksc.cdn.model.domain.domaindetail;
 
 
 import com.ksc.cdn.KscClientException;
