+++ ksc-sdk-java-cdn/src/main/java/com/ksc/cdn/model/domain/domaindetail/TestUrlConfig.java
@@ -1,0 +1,19 @@
+package com.ksc.cdn.model.domain.domaindetail;
+
+/**
+ * TestUrlConfig
+ *
+ * @author jiangran@kingsoft.com
+ * @date 2016/11/15
+ */
+public class TestUrlConfig{
+    private String TestUrl;
+
+    public String getTestUrl() {
+        return TestUrl;
+    }
+
+    public void setTestUrl(String TestUrl) {
+        TestUrl = TestUrl;
+    }
+}
