+++ ksc-sdk-java-cdn/src/main/java/com/ksc/cdn/model/domain/domaindetail/VideoSeekConfig.java
@@ -1,0 +1,19 @@
+package com.ksc.cdn.model.domain.domaindetail;
+
+/**
+ * VideoSeekConfig
+ *
+ * @author jiangran@kingsoft.com
+ * @date 2016/11/15
+ */
+public class VideoSeekConfig{
+    private String Enable;
+
+    public String getEnable() {
+        return Enable;
+    }
+
+    public void setEnable(String enable) {
+        Enable = enable;
+    }
+}
