+++ ksc-sdk-java-cdn/src/main/java/com/ksc/cdn/model/statistic/flow/FlowResult.java
@@ -1,0 +1,18 @@
+package com.ksc.cdn.model.statistic.flow;
+
+import com.ksc.cdn.model.statistic.CommonFieldResult;
+
+/**
+ * Created by jiangran on 19/10/2016.
+ */
+public class FlowResult extends CommonFieldResult {
+    private FlowDataByTime[] Datas;
+
+    public FlowDataByTime[] getDatas() {
+        return Datas;
+    }
+
+    public void setDatas(FlowDataByTime[] datas) {
+        Datas = datas;
+    }
+}
