+++ ksc-sdk-java-cdn/src/main/java/com/ksc/cdn/model/statistic/hitrate/HitRateRequest.java
@@ -1,4 +1,4 @@
-package com.ksc.cdn.model.statistic;
+package com.ksc.cdn.model.statistic.hitrate;
 
 import com.ksc.cdn.KscClientException;
 import com.ksc.cdn.model.valid.CommonValidUtil;
