+++ ksc-sdk-java-cdn/src/main/java/com/ksc/cdn/model/statistic/pv/PVResult.java
@@ -1,0 +1,26 @@
+package com.ksc.cdn.model.statistic.pv;
+
+
+import com.ksc.cdn.model.statistic.CommonFieldResult;
+
+import java.util.List;
+
+/**
+ * Created by CrazyHorse on 17/10/2016.
+ */
+public class PVResult extends CommonFieldResult {
+
+    private PvDataByTime[] Datas;
+
+    public PvDataByTime[] getDatas() {
+        return Datas;
+    }
+
+    public void setDatas(PvDataByTime[] datas) {
+        Datas = datas;
+    }
+
+
+
+
+}
