+++ ksc-sdk-java-cdn/src/test/java/com/ksc/cdn/test/StatisticTest.java
@@ -4,6 +4,12 @@ import com.ksc.cdn.KscCdnClient;
 import com.ksc.cdn.KscCdnStatistics;
 import com.ksc.cdn.model.enums.*;
 import com.ksc.cdn.model.statistic.*;
+import com.ksc.cdn.model.statistic.bandwidth.BpsResult;
+import com.ksc.cdn.model.statistic.flow.FlowResult;
+import com.ksc.cdn.model.statistic.hitrate.HitRateDetailResult;
+import com.ksc.cdn.model.statistic.hitrate.HitRateRequest;
+import com.ksc.cdn.model.statistic.hitrate.HitRateResult;
+import com.ksc.cdn.model.statistic.pv.PVResult;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
